Update user controller and related files.

+++ LineItem.Repositories/Interfaces/IUserRepository.cs
@@ -8,9 +8,9 @@ public interface IUserRepository
 {
     public Task<long> CreateAsync(UserModel user, CancellationToken cancellationToken);
 
-    public Task<UserModel?> RetrieveByIdAsync(int id, CancellationToken cancellationToken);
+    public Task<UserModel?> RetrieveByIdAsync(long id, CancellationToken cancellationToken);
 
     public Task<bool> UpdateAsync(UserModel user, CancellationToken cancellationToken);
 
-    public Task<bool> DeleteAsync(int id, CancellationToken cancellationToken);
+    public Task<bool> DeleteAsync(long id, CancellationToken cancellationToken);
 }
