+++ LineItem.Services/Interfaces/IUserService.cs
@@ -8,12 +8,12 @@ public interface IUserService
 {
     public Task<long> CreateAsync(UserModel user, CancellationToken cancellationToken);
 
-    public Task<UserModel?> RetrieveByIdAsync(int id, CancellationToken cancellationToken);
+    public Task<UserModel?> RetrieveByIdAsync(long id, CancellationToken cancellationToken);
 
     public Task<bool> UpdateAsync(
         UserModel user,
         CancellationToken cancellationToken
     );
 
-    public Task<bool> DeleteAsync(int id, CancellationToken cancellationToken);
+    public Task<bool> DeleteAsync(long id, CancellationToken cancellationToken);
 }
